Move the gui to compute recharge to its own module
